requirements.txt
ijson==3.2.2
lf-metric-emitter==0.0.3
requests==2.31.0